Split manager and add api tests
