+++ Testing/Persister.php
@@ -4,6 +4,7 @@ namespace UJM\ExoBundle\Testing;
 
 use Claroline\CoreBundle\Entity\Resource\ResourceNode;
 use Claroline\CoreBundle\Entity\Resource\ResourceType;
+use Claroline\CoreBundle\Entity\Role;
 use Claroline\CoreBundle\Entity\User;
 use Claroline\CoreBundle\Entity\Workspace\Workspace;
 use Claroline\CoreBundle\Persistence\ObjectManager;
@@ -11,9 +12,13 @@ use UJM\ExoBundle\Entity\Category;
 use UJM\ExoBundle\Entity\Choice;
 use UJM\ExoBundle\Entity\Exercise;
 use UJM\ExoBundle\Entity\ExerciseQuestion;
+use UJM\ExoBundle\Entity\Hint;
 use UJM\ExoBundle\Entity\InteractionOpen;
 use UJM\ExoBundle\Entity\InteractionQCM;
+use UJM\ExoBundle\Entity\Paper;
 use UJM\ExoBundle\Entity\Question;
+use UJM\ExoBundle\Entity\TypeQCM;
+use UJM\ExoBundle\Manager\PaperManager;
 
 /**
  * Simple testing utility allowing to create and persist
@@ -31,9 +36,25 @@ class Persister
      */
     private $exoType;
 
-    public function __construct(ObjectManager $om)
+    /**
+     * @var Role
+     */
+    private $userRole;
+
+    /**
+     * @var TypeQCM
+     */
+    private $multipleChoiceType;
+
+    /**
+     * @var PaperManager
+     */
+    private $paperManager;
+
+    public function __construct(ObjectManager $om, PaperManager $paperManager)
     {
         $this->om = $om;
+        $this->paperManager = $paperManager;
     }
 
     /**
@@ -62,12 +83,21 @@ class Persister
         $question->setTitle($title);
         $question->setInvite('Invite...');
 
+        if (!$this->multipleChoiceType) {
+            $this->multipleChoiceType = new TypeQCM();
+            $this->multipleChoiceType->setCode(1);
+            $this->multipleChoiceType->setValue('Multiple response');
+            $this->om->persist($this->multipleChoiceType);
+        }
+
         $interactionQcm = new InteractionQCM();
         $interactionQcm->setQuestion($question);
+        $interactionQcm->setTypeQCM($this->multipleChoiceType);
+        $interactionQcm->setWeightResponse(true);
 
         for ($i = 0, $max = count($choices); $i < $max; ++$i) {
-            $choices[$i]->setInteractionQCM($interactionQcm);
             $choices[$i]->setOrdre($i);
+            $interactionQcm->addChoice($choices[$i]);
         }
 
         $this->om->persist($interactionQcm);
@@ -125,6 +155,7 @@ class Persister
             $node->setResourceType($this->exoType);
             $node->setWorkspace($user->getPersonalWorkspace());
             $node->setClass('UJM\ExoBundle\Entity\Exercise');
+            $node->setGuid(time());
             $exercise->setResourceNode($node);
             $this->om->persist($node);
         }
@@ -132,6 +163,16 @@ class Persister
         $this->om->persist($exercise);
 
         return $exercise;
+    }
+
+    /**
+     * @param User      $user
+     * @param Exercise  $exercise
+     * @return Paper
+     */
+    public function paper(User $user, Exercise $exercise)
+    {
+        return $this->paperManager->createPaper($user, $exercise);
     }
 
     /**
@@ -148,6 +189,13 @@ class Persister
         $user->setMail($username . '@mail.com');
         $user->setGuid($username);
         $this->om->persist($user);
+
+        if (!$this->userRole) {
+            $this->userRole = $this->role('ROLE_USER');
+            $this->om->persist($this->userRole);
+        }
+
+        $user->addRole($this->userRole);
 
         $workspace = new Workspace();
         $workspace->setName($username);
@@ -172,5 +220,30 @@ class Persister
         $this->om->persist($category);
 
         return $category;
+    }
+
+    /**
+     * @param string $name
+     * @return Role
+     */
+    public function role($name)
+    {
+        $role = new Role();
+        $role->setName($name);
+        $role->setTranslationKey($name);
+        $this->om->persist($role);
+
+        return $role;
+    }
+
+    public function hint(Question $question, $text, $penalty = 1)
+    {
+        $hint = new Hint();
+        $hint->setValue($text);
+        $hint->setPenalty($penalty);
+        $hint->setQuestion($question);
+        $this->om->persist($hint);
+
+        return $hint;
     }
 }
